third_party/WebKit/LayoutTests/fast/events/event-listener-on-attribute-inside-shadow-dom-expected.txt

This tests ensures that an event listener on an attribute node inside a shadow DOM is properly unregistered when parent element of the attribute is moved to a new document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.internals.touchEventHandlerCount(document) is 1
PASS window.internals.touchEventHandlerCount(document) is 0
PASS successfullyParsed is true

TEST COMPLETE
